refactor: refatoração geral na utilização de MethodResponse no projeto.

+++ src/main/java/com/camp/manager/domain/entity/utils/MethodResponse.java
@@ -2,7 +2,7 @@ package com.camp.manager.domain.entity.utils;
 
 
 public record MethodResponse<T>(
-        Long status,
+        int status,
         String message,
         T data) {
 }
